fix: remove readability requirement when setting property
`setTypeInstance` is not supposed to require the given relationship to
be readable.

By removing the requirement it is now possible to set a relationship
that is not readable.

The method is also used when checking if a property is readable,
which until now threw an `TypeRetrievalAccessException` in case it is
not. Now the more fitting `PropertyAccessException` will be thrown
instead.