feature: skip prefixing of context bound paths
Prefixing a context-bound paths does not seem
to be a reasonable thing to do, as it would
either mean that the path was previously invalid
or becomes invalid due to the prefixing in almost
all use-cases. For now paths bound to a specific
context are silently ignored during the prefixing
of multiple paths. This seems to be the most
reasonable thing to do, as context-bound paths
may be easily mixed with paths without context
and prefixing the paths without context is
completely valid.