feature: prevent invalid to-many usage in DQL
Sorting by columns in to-many relationships makes
logically no sense and implies a mistake instead
of intended usage. Thus, paths containing to-many
relationships are no longer allowed when converting
sort methods into DQL.

The approach allows to prevent the usage of such
paths in conditions and selects as well. Especially
`InvertedBoolean` is such candidate, as it does
behave non-intuitively for to-many relationships.

For PHP evaluations this limitation is not supported yet.