refactor: remove additional format/type validation
In cases where the input types and array formats are already
type hinted it is not worth it to mix validation inside the
processing logic. Instead, the input can be validated before
passing it into the methods processing it.

For Drupal filters this is already done when using the
`listFromRequest` method. For JSON:API update/create request
body data it needs to be done manually.