feature: rework resource property configuration
Instead of `getProperties`, `AbstractResourceType` now invokes
`configureProperties`, which allows the configuration of
properties in a more guided way with more features and better
type-hinting.

Implementing `getUpdatableProperties(object)` is no longer
necessary, as `configureProperties` can be used instead.

Internally the property configuration is passed around
inside the engine and replaces classes like `Property`
and `PropertyDefinition`.

The new config classes also carry additional information,
like if a relationship has a to-one or to-many cardinality,
which is used to validate values read from entities instead
of trusting the property read from the entity to determine
the cardinality.

There were other adjustments as well, the list may be
incomplete:

* `PropertyReader` does no longer accept `null` values,
  callers need to check for `null` themselves
* the caching done in `AttributeTypeReslover` was reduced
* added related and unrelated TODO markers
* reduced caching done in `CachingResourceType`
* prefer `instanceof` over `is_a`