feature: restore classes and update changelog
To allow for a smother transition to the new property
configuration, the old `PropertyBuilder` and methods
to create it were restored with some adjustments but
marked as deprecated.

Some other methods in the `AbstractResourceType` were
restored/adjusted as well. Also, by moving some code
in the `AbstractResourceType` the line-diff to the
previous property configuration approach was reduced.

The changelog was updated accordingly and freed from
information, that did not seem relevant for projects
using this library.