fix: Handle git CVE-2022-24765
Git added mechanisms to prevent fraudulent directory and file manipulations. However, we kind of need to perform some odd-looking actions during the repository split. Toward that end, the build-directory is defined as a git file ceiling, even though it is not in the user's path.

For more information, see https://github.blog/2022-04-12-git-security-vulnerability-announced/.