fix: consider `null` when processing paths with relationships
The object assertion was added because we must not try to
process additional path segments when we reached a non-object
as target. However, `null` is the exception, as
`getValueByPropertyPath` will already consider it and
abort by returning `null` as final value. Hence, we
must check for it before asserting the new target.