fix: handle required properties correctly in validation
Due to the complexity of using the same constructor
parameter for options and fields, Symfony Validators
version 5.4.34 broke the initialization of
`Assert\Collection` with an empty list of fields.

This commit mitigates this problem and also fixes
additional ones in EDT, that were discovered during
testing. Specifically constraints of the `attributes`
and ``relationships` fields were not executed, if
if they were not present in the request at all.