refactor(property config)!: improve approaches
Continue to solve the concerns in the referenced issue.
Replacement of all behavior factories with static `createFactory`
methods in the corresponding behavior implementation
was not possible, because their parent interfaces require
typehinting, that can not be done using anonymous classes.

Maybe after the removal of `TCondition` and `TSorting` it
becomes possible to remove additional factories, but
other will need to remain.

Refs: #115