refactor(property config): re-add old methods
To reduce compatibility breaking changes, the
previously replaced method was added again,
but marked as deprecated.

A template parameter was adjusted as well. Neither
having it defined in the method nor having it
defined in the class docblock seem correct. This
needs to be investigated further for all behavior
factory classes, but for now the (understandable)
concern was added to the phpstan baseline of
concerns.

Refs: #115