feat(property config): avoid recursion loop
Passing and using a config builder as relationship
type (as shown in the README.adoc) does only work
if we postpone actually
requiring the relationship type. E.g. if
the config builder for `Book` resources references
authors and the authors have a reference to the books,
then on building the `Book` config builder it needs the
`Author` relationship type from the `Author`
config builder, so the author config needs to be build too,
which in turn requires the `Book` config builder to be
build first.

By not requiring the actual relationship resource type
when building, but a provider instead, all the configs
can be build first and only when the relationship type
is actually needed for some task it is requested from
the provider.

Keeping the possibility to initialize relationship
config builders from the actual relationship still
works outside of the context of resource config
builders, e.g. when only using schema config
builders.

Refs: #115