refactor(request handling): move methods
Having methods that are tailored for specific
kind of request in the `RequestTransformer`
carries the danger of making that class some mishmash
of helper methods without clear scope. Instead,
logic specific for updates and creations should
reside in the `CreationRequest` and `UpdateRequest`
classes, even though it makes the logic harder to
re-use.