feature(request handling): add Fractal initializer
To convert the resources created by request classes
(e.g. `Item` and `Collection`) into Symfony `Response`
instances, the library already made use of Fractal
transformer classes. However, to actually use the
transformers, a Fractal `Manager` instance is needed,
whose creation is non-trivial.

This commit adds the logic to create such `Manager`
with reasonable defaults. To do so, a `IncludeValidator`
is needed as well.