feature(filter transformation)!: allow better validation
Introduce `ConditionInterface` implementations, which
allow to transform a Drupal condition into a condition
instances, based on their condition factory.

These implementations will able to provide information
to validate their schema. However, this is not yet
fully implemented.

As it is needed for "always true" and "always false"
conditions/operators, the `path` field in Drupal
filters are no longer required unconditionally.
Most individual conditions will still require it,
but the Drupal filter schema itself allows it to
be missing, because there is no sensible value for
"always true" and "always false".

This change also includes the movement of the `getSupportedOperators`
method into its own interface and the removal of arbitrary
input support. I.e. the input must be a Drupal condition in
array format.